feat: deprecate htmlHelper mixin
